requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
dnspython==2.5.0
python-dotenv==1.0.1
pytest==8.0.2
firebase-admin==6.4.0
flask==3.0.2
selenium==4.18.1
webdriver-manager==4.0.1
fake-useragent==1.4.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0 